This fix has a bunch of bug fixes.
Sorry for clubbing them all together into a single commit.
First, the upstream serialization fmt::Display does not work as expected
for the purposes for elements-miniscript. It serializes the explicit
asset and explicit values as 32 bytes and u64 respectively.
Second is a bug that is specific to elements-miniscript, we were
not checking that the correct possition and parent are being used while
creating a new miniscript. This allowed to create fragments that allowed
to use confidential::Value in asset_eq fragments.

Third, this removes the from_str implementations from AssetExpr,
ValueExpr and SpkExpr. These expressions cannot be independent as they
are parent context dependent. Add a from_arg_str implementation instead