Add parsing from hashed version of scripts
This required in order to deal with legacy scripts while parsing from
scripts. When creating scripts, we normally create them using the
regular Script variant. But when parsing from script, we store the
hashed version for legacy scripts as we cannot compute the pre-image.

This is helpful in psbt workflow where we need to infer miniscript from
Script