[MOVE ONLY] covenants: move ExtParams to its own module
Right now ExtParams and Ext are together in one module, and are confused
in places. My goal here is to isolate ExtParams within the extension module
as much as possible and have users mostly just deal with Ext.

Of course, to convert between CovenantExt<A> and CovenantExt<B> they will
need to deal with ExtParam, but we will expose that in a narrow sense
rather than having Descriptor<ExtParam> and a "TranslateExt" trait that
actually manipulates ExtParams.

Re-exports things in their own locations to minimize disruption to the
rest of the crate. At the end of this PR we will fix all the export paths.