refactor `Descriptor` to be parameterized by Extension, not ExtParam
It was confusing to have `Descriptor` parameterized by `ExtParam`;
it forced all descriptors to have the `CovenantExt` extension which
will prevent any alternate extensions (eg Simplicity) and also made
the `Extension` trait redundant.

It also made the `NoExt` struct useless even though you were able to
construct a `Descriptor<Pk, NoExt>` and do one or two things with it.
This is very confusing.

There was also confusion in the codebase in multiple places where
extensions and extension parameters were confused. This commit introduces
the ExtParamTranslator and TranslateExtParam traits to handle translation
of the parameters; then it introduces some helper implementations to make
this more ergonomic.

It should now be actually possible to use descriptors without any
extensions, which will facilitate Simplicity, infalible conversions to
rust-miniscript structures, and hopefully let people figure out how to
use the traits in this library without immediately being swallowed by
an impossible maze of traits.

BTW sorry for the huge diff on this commit. Everything was intertwined
and I had to change everything at once to get it to compile. It should
be less intertwined now so that future changes won't be so bad.