Add price_oracle1 and price_oracle1_w
Arith becomes ArithInner. ArithInner does not check any rules on
placement on price_oracle1 or price_oracle1_w. Arith is the checked
version that we use the API. Even though it is ugly for the user to
write corresponding oracle1 and oracle1_w versions, atleast the type
system will test if it is incorrect