expression: stop special-casing the text `slip77`
In a previous incarnation of CT descriptors we considered "musig" and
"slip77" expressions to constitute keys, allowing them to be parsed in
contexts where normally only keys would be allowed (and _not_ expanding
them as expression trees which would then need to be reconstructed).

This is still plausibly the direction we want to go in for MuSig keys,
but it never made a lot of sense for SLIP77, and indeed, we now want to
go a different direction, where SLIP77 expressions are allowed only as
the first argument of a CT descriptor.