tr: Add Simplicity leaf to TapTree
Extends the tap iterator to handle both kinds of leaves.
The iterator returns an enum that has to be split into either variant
as required. For now all methods assume Miniscript and will panic if
Simplicity is encountered. We will add Simplicity support to these
methods in the next commits.