simplicity: Parse policy
We implement the FromTree trait for Policy to do the heavy lifting.
This was policy/embed.rs in rust-simplicity:
https://github.com/BlockstreamResearch/rust-simplicity/commit/35d3fca17d0733ca12f0f1726c453b5071486220

I would like to implement FromStr for Policy, but the orphan rule won't
let me. As a workaround I implement FromStr for a crate-local wrapper of
Policy. Users will never encounter this type.