tr: Spend Simplicity descriptors
TapLeafScript is extended to handle everything that is required for
spending in a generic way. The witness for Simplicity looks familiar:
"Witness elements", "witness script" and the control block.
What these are is slighly different:

There is one "Witness element" which is one large byte blob.
This deserializes to the Simplicity program plus its witness data.
Unlike Miniscript, there is always exactly one witness element.

"Witness script" is a Bitcoin Script that contains the CMR.

The control block is as usual. Even though there are these differences,
the code for computing the "witness size" should be general enough to
work for both Miniscript and Simplicity.