tr: Weight calculations for Simplicity
The length of the witness stack is always the same: 3 elements.
There is the serialized program+witness, the CMR and the control block.
The maximum witness size is trickier, because of the custom encoding
that Simplicity uses and because of the existence of disconnect (will be
present in asm fragments). There exists no function to compute a bound,
and the real witness size would be much smaller than what this function
returned. max_satisfaction_size will return an error for Simplicity
leaves. We can still choose the best leaf to spend because the spending
methods generate a witness from a satisfier, which we _can_ do.