CT descriptor: remove generic from blinding key
The generic was only used by the `Bare` variant, which was
inconsistent with its corresponding "secret" variant, `View`.

The generic could be used to handle the conversion from
`ConfidentialDescriptor<DescriptorPublicKey>` to
`Descriptor<DefiniteDescriptorKey>`, specifically making sure
that the descriptor blinding key does not have wildcards.

For the `Bare` variant this could be done, but this cannot
happen for the `View` variant, since we don't have the definite
version of `DescriptorSecretKey`.