Remove type_check from Property trait
The blanket implementation of `type_check` is only used on `CompilerExtData`
while `ExtData` and `Type` override the impl and don't need the `child`
parameter. Moreover, the fn isn't called as Property generic.

So the blanket implementation of `type_check` is moved as a impl in
`CompilerExtData` and the overrides in `ExtData` and `Type` are moved
as simple impl on the type, making it possible to remove  the unused `child`
parameter.