doc: improve list of fields covered by sighash
This will result in the rendered numbering being different from the
numbering in the source code, but that's probably better than the
existing situation where the "3b" item does not show up as an
independent list item.

There's no good way I can see to fix this because of limitations of
rustfmt's list facilities. And anyway this is legacy code. So just do
something that makes clippy happy.