ci: copy workflow and contrib/ scripts from rust-elements
This introduces the rust-bitcoin-maintainer-tools CI setup, which ensures
that all tool versions are pinned, tests all examples, etc. I had to tweak
a couple things (replacing the list of features, renaming elementsd-tests
to bitcoind-tests) but otherwise this was a straight copy/paste.

We also update all our CI actions to v4 from the now-deleted v2.