clippy: introduce nightly-version and fix all clippy issues with it
Copy nightly-version from rust-elements. The next commit will use it in
Github CI. I won't bother making a cronjob for it because this crate is
so unused that it doesn't seem worthwhile opening constant PRs for CI
updates. We will try to do this manually every several months, or add
a cronjob if/when this becomes used.

I added an #[allow(unused)] tag on the elementsd test crate. Maybe this
trait is actually unused and should be deleted. But the elementsd tests
are old and bitrotted and I'd prefer to avoid touching them for purposes
of this PR.

This commit also introduces a lockfile. Basically, from this commit
onward I want my local CI to pass, so I am throwing in several things
at once.